+++ js/form.js
@@ -14,7 +14,6 @@ const offerTitleInput = offerForm.querySelector('#title');
 const resetButton = offerForm.querySelector('.ad-form__reset');
 const main = document.querySelector('main');
 
-const ALERT_SHOW_TIME = 5000;
 const MIN_TITLE_LENGTH = 30;
 const MAX_TITLE_LENGTH = 100;
 const DEFAULT_MAX_PRICE = 1000000;
@@ -31,6 +30,11 @@ const numberOfGuests = {
   2: ['1', '2'],
   3: ['1', '2', '3'],
   100: ['0'],
+};
+
+const Keys = {
+  ESCAPE: 'Escape',
+  ESC: 'Esc',
 };
 
 // Синхронизация полей времени заезда и выезда
@@ -81,17 +85,10 @@ offerTitleInput.addEventListener('input', () => {
 
 
 // Валидация количества гостей и комнат
-
 const validateRooms = () => {
   const roomValue = selectRooms.value;
   capacityOptions.forEach((option) => {
-
-    // Что проверяет это условие:
-    // Проверяем наличие option.value (кол-во гостей) в значении свойства объекта numberOfGuests. Значения у этого объекта являются массивами.
-    // Используя метод indexOf мы получаем индекс элемента массива (проверяем наличие элемента в массиве).
-    // Если indexOf возвращает -1 , то это значит, что элемента в массиве нет и для соответствующего <option> ставим атрибут disabled="true". Если элемент есть в массиве, то индекс будет >=0, и соответственно disabled="false".
     let isDisabled = (numberOfGuests[roomValue].indexOf(option.value) === -1);
-
     option.selected = numberOfGuests[roomValue][0] === option.value;
     option.disabled = isDisabled;
     option.hidden = isDisabled;
@@ -106,31 +103,10 @@ validateRooms();
 
 selectRooms.addEventListener('change', onSelectRoomsChange);
 
-// Сообщение об ошибке
-const showAlert = (message) => {
-  const alertContainer = document.createElement('div');
-  alertContainer.style.zIndex = 100;
-  alertContainer.style.position = 'absolute';
-  alertContainer.style.left = 0;
-  alertContainer.style.top = 0;
-  alertContainer.style.right = 0;
-  alertContainer.style.padding = '10px 3px';
-  alertContainer.style.fontSize = '20px';
-  alertContainer.style.textAlign = 'center';
-  alertContainer.style.backgroundColor = 'red';
-
-  alertContainer.textContent = message;
-
-  document.body.append(alertContainer);
-
-  setTimeout(() => {
-    alertContainer.remove();
-  }, ALERT_SHOW_TIME);
-}
 
 const closeMessage = (message) => {
   document.addEventListener('keydown', (evt) => {
-    if (evt.key === 'Escape') {
+    if (evt.key === Keys.ESCAPE || evt.key === Keys.ESC) {
       message.remove();
     }
   });
@@ -161,26 +137,21 @@ const showErrorMessage = () => {
   closeMessage(errorMessage);
 };
 
-
 const setUserFormSubmit = () => {
   offerForm.addEventListener('submit', (evt) => {
     evt.preventDefault();
 
-    sendData(
-      () => showSuccessMessage(),
-      () => showErrorMessage(),
-      new FormData(evt.target),
-    );
+    sendData(showSuccessMessage, showErrorMessage, new FormData(evt.target));
   });
 };
 
+setUserFormSubmit();
+
+
 resetButton.addEventListener('click', (evt) => {
   evt.preventDefault();
-
   offerForm.reset();
   setDefaultAdress();
   resetMainPinMarker();
   onOfferTypeChange();
 })
-
-export {setUserFormSubmit, showAlert};
