+++ js/make-elements.js
@@ -1,8 +1,14 @@
-import { TYPES } from './data.js';
 import { numDecline } from './util.js';
 
 const cardTemplate = document.querySelector('#card').content;
 const popupTemplate = cardTemplate.querySelector('.popup');
+
+const TYPES = {
+  palace: 'Дворец',
+  flat: 'Квартира',
+  house: 'Дом',
+  bungalow: 'Бунгало',
+};
 
 const renderFeatures = (features, container) => {
   let list = container.querySelectorAll('li');
