+++ js/util.js
@@ -1,27 +1,3 @@
-const getRandomNumber = (min, max, digit = 0) => {
-  if (min < 0 || max < 0) {
-    return -1;
-  }
-
-  if (max < min) {
-    [min, max] = [max, min]
-  }
-
-  return digit === 0 ? Math.floor(Math.random() * (max - min + 1)) + min : (Math.random() * (max - min) + min).toFixed(digit);
-};
-
-const getRandomArrayElement = (elements) => {
-  return elements[getRandomNumber(0, elements.length - 1)];
-};
-
-const shuffleArray = (array) => {
-  for (let i = array.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    [array[i], array[j]] = [array[j], array[i]];
-  }
-  return array;
-};
-
 const numDecline = (num, nominative, genitiveSingular, genitivePlural) => {
   if(num > 10 && (Math.round((num % 100) / 10)) == 1) {
     return genitivePlural;
@@ -36,4 +12,4 @@ const numDecline = (num, nominative, genitiveSingular, genitivePlural) => {
   }
 };
 
-export {getRandomNumber, getRandomArrayElement, shuffleArray, numDecline};
+export {numDecline};
